requirements.txt
flask
requests
prometheus_client
python-dotenv
gunicorn==21.2.0